Implemented constraint satisfaction solver
This version of the solver breaks down if any value has all of it's possible values eliminated. i.e. there is no form of backtracking or other approach.

The performance of the `Solver:prune_next_arc()` leaves much room for improvement.

Can successfully solve a sudoku with only one solution.